fix(draft-capital): lazy-load trade simulator to fix /league bundle budget
Inlining the simulator pushed the /league page bundle to 174.1 KB,
9.1 KB over its 165 KB budget. Extract it to ./_trade-simulator.jsx
and import via next/dynamic({ ssr: false }) so its chunk loads on
demand instead of joining the page-specific bundle. /league/page is
back to 162.5 KB.
